requirements.txt
python-telegram-bot==21.5
pandas==2.2.1